--- src/assets/content/camera-share-guides/macos/edge.md
@@ -1,37 +1,0 @@
-### For Edge Users
-
-#### Step 1:
-In the top-left search bar of your browser, click the settings icon and tap on site settings
-
-
-![edge-site-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/816/original/edge_camera_share_s1.png?1733907413)
-
-
-
-#### Step 2:
-Locate the Camera option and set it to Allow
-
-
-![edge-site-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/817/original/edge_camera_share_s1.png?1733907458)
-
---
-
-#### Still facing problem??
-Please follow the steps mentioned below to enable system-level permissions
-
-
-### System Level Fixes
-
-#### Step 1:
-Select the Apple menu and select System Settings
-
-
-![mac-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/700/original/camera_system_step1.png?1733846308)
-
-
-
-#### Step 2:
-Select Privacy & Security in the sidebar. Select Camera on the right and enable access to camera for your browser in the list
-
-
-![mac-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/701/original/camera_system_step2.png?1733846340)
