--- src/assets/content/camera-share-guides/macos/safari.md
@@ -1,53 +1,0 @@
-### For Safari Users
-
-#### Step 1:
-In the top-left search bar of your browser, click the settings icon and tap on site settings
-
-
-![safari-site-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/818/original/safari_camera_share_s1.png?1733907483)
-
-
-
-#### Step 2:
-Locate the Camera option and set it to Allow
-
-
-![safari-site-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/819/original/safari_camera_share_s2.png?1733907506)
-
---
-
-#### Still facing problem??
-Please follow the steps mentioned below to enable system-level permissions
-
-
-### System Level Fixes
-
-#### Step 1:
-Select the Apple menu and select System Settings
-
-
-![safari-site-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/700/original/camera_system_step1.png?1733846308)
-
-
-
-#### Step 2:
-Select Privacy & Security in the sidebar. Select Camera on the right and enable access to camera for your browser in the list
-
-
-![safari-site-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/701/original/camera_system_step2.png?1733846340)
-
----
-
-### Troubleshooting
-
-#### Camera Access
-   - Check System Preferences
-   - Verify Safari settings
-   - Reset permissions
-   - Update macOS
-
-#### Video Quality
-   - Check internet connection
-   - Verify camera settings
-   - Update Safari
-   - Reset camera
