--- src/assets/content/camera-share-guides/windows/edge.md
@@ -1,37 +1,0 @@
-### For Edge Users
-
-#### Step 1:
-In the top-left search bar of your browser, click the settings icon and tap on site settings
-
-
-![safari-site-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/816/original/edge_camera_share_s1.png?1733907413)
-
-
-
-#### Step 2:
-Locate the Camera option and set it to Allow
-
-
-![safari-site-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/817/original/edge_camera_share_s1.png?1733907458)
-
---
-
-#### Still facing problem??
-Please follow the steps mentioned below to enable system-level permissions
-
-
-### System Level Fixes
-
-#### Step 1:
-Select the Apple menu and select System Settings
-
-
-![windows-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/823/original/window_camera_share_s1.png?1733911602)
-
-
-
-#### Step 2:
-Select Privacy & Security in the sidebar. Select Camera on the right and enable access to camera for your browser in the list
-
-
-![windows-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/824/original/window_camera_share_s2.png?1733911631)
