--- src/assets/content/camera-share-guides/windows/firefox.md
@@ -1,36 +1,0 @@
-### For Firefox Users
-
-#### Step 1:
-In the top-left search bar of your browser, click the settings icon and tap on site settings
-
-
-![safari-site-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/698/original/camera_safari_step1.png?1733846008)
-
-
-
-#### Step 2:
-Locate the Camera option and set it to Allow
-
-
-![safari-site-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/699/original/camera_safari_step2.png?1733846169)
-
---
-
-#### Still facing problem??
-Please follow the steps mentioned below to enable system-level permissions
-
-### System Level Fixes
-
-#### Step 1:
-Select the Apple menu and select System Settings
-
-
-![windows-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/823/original/window_camera_share_s1.png?1733911602)
-
-
-
-#### Step 2:
-Select Privacy & Security in the sidebar. Select Camera on the right and enable access to camera for your browser in the list
-
-
-![windows-settings](https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/099/824/original/window_camera_share_s2.png?1733911631)
